pyproject.toml
[project]
name = "lastfm-global-trends"
version = "0.1.0"
description = "Shiny for Python dashboard exploring global music trends via the Last.fm API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cachetools>=7.0.5",
    "duckdb>=1.5.0",
    "itables>=2.7.1",
    "jinja2>=3.1.6",
    "kaggle>=2.0.0",
    "pandas>=3.0.1",
    "plotly>=6.3.1",
    "pylast==7.0.2",  # pinned: raw_request() uses pylast._Request (private API) — verify on upgrade
    "python-dotenv>=1.2.2",
    "shiny>=1.5.1",
    "shinywidgets>=0.7.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "rsconnect-python>=1.28.2",
    "ruff>=0.15.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade – modernize syntax (e.g. f-strings, type hints)
    "B",   # flake8-bugbear – catches real bugs, not style pedantry
    "C4",  # flake8-comprehensions – cleaner list/dict/set expressions
    "SIM", # flake8-simplify – simplify boolean checks, redundant code
]
ignore = [
    "E501",   # line too long – ruff format handles this
    "SIM108", # ternary operator – sometimes less readable
]
